Update weapon_bonus.py (#6)
* Update weapon_bonus.py

Created a dictionary (weapon_skill_bonuses) to store the bonuses for each weapon and corresponding weapon skill.
Each weapon is a key in the dictionary, and the value is another dictionary mapping weapon skills to their bonuses.
Using dictionaries for lookups is faster and more efficient than multiple if-elif statements.
The code is shorter and clearer, making it easier to understand and maintain.

* Update weapon_bonus.py

Noted Vajra was used for both Mandallic Stab and Pyrrhic Kleos. Added Terpsichore to dictionary and assigned to Pyrrhic Kleos.

* Update weapon_bonus.py

Updated Heishi to Heishi Shorinken, removed additional Terpsichore line.