Adjust how algorithm initiates set building.
Previously: Initialize the best set by unequipping gear in all slots to be tested.

Now: If only one item is selected in a slot: equip that item. If two or more are selected: initialize with an empty slot. Main and sub slots are treated separately by randomly equipping one of the selected items.